move all of the plaid link logic to a component
