+++ client/src/pages/Dashboard.tsx
@@ -7,17 +7,13 @@ import { apolloClient } from '..';
 import { Button, Chip, Dialog, DialogBody, DialogFooter, DialogHeader, Tooltip } from '@material-tailwind/react';
 import { PlaidLinkOptions, usePlaidLink } from 'react-plaid-link';
 import { Loading } from '@/components/Loading';
-import { PlaidLinkButton } from '@/components/Plaid/PlaidLinkButton';
+import { PlaidLinkPrompt } from '@/components/Plaid/PlaidLinkPrompt';
 
 
 export function Dashboard({ }: any) {
 
   const { data: isBankLinked, refetch: refetchIsBankLinked } = useQuery<any>(IS_BANKACCOUNT_LINKED);
-  const [getLinkToken, { data: linkTokenData }] = useLazyQuery<any>(PLAID_CREATE_LINK_TOKEN, {
-    fetchPolicy: "no-cache"
-  });
-  const [linkToken, setLinkToken] = useState<string | null>(null);
-  const [plaidLinkReady, setPlaidLinkReady] = useState<boolean>(false);
+  const [openPlaidPrompt, setOpenPlaidPrompt] = useState<boolean>(false);
   const [unlinkBankAccount] = useMutation<any>(PLAID_UNLINK_BANK, {
     refetchQueries: [
       { query: IS_BANKACCOUNT_LINKED },
@@ -31,8 +27,6 @@ export function Dashboard({ }: any) {
   const { data: transactionsData, loading: loadingTransactions } = useQuery<any>(PLAID_GET_TRANSACTIONS);
 
   const [totalBalance, setTotalBalance] = useState<number | undefined>(undefined);
-  const [selectedProducts, setSelectedProducts] = useState<string[]>([]);
-  const [openProductSelection, setOpenProductSelection] = useState<boolean>(false);
 
   const currencyFormatter = new Intl.NumberFormat('en-US', { style: 'currency', currency: 'USD' });
 
@@ -57,20 +51,11 @@ export function Dashboard({ }: any) {
     }
   }, [balanceData, transactionsData, accountData]);
 
-  const handleClickProduct = (product: string) => {
-    if (selectedProducts.includes(product)) {
-      const index = selectedProducts.indexOf(product);
-      setSelectedProducts(selectedProducts.filter(item => item !== product));
-    } else {
-      setSelectedProducts(selectedProducts.concat(product));
-    }
-  }
-
   return (
     <div className="ml-24 flex flex-col min-h-screen min-w-0 max-w-full overflow-hidden">
       {!isBankLinked?.bankAccountLinked ?
         <div className='grow flex justify-center items-center'>
-          <Button onClick={() => setOpenProductSelection(!openProductSelection)} variant="gradient" size="lg">
+          <Button onClick={() => setOpenPlaidPrompt(!openPlaidPrompt)} variant="gradient" size="lg">
             <div className="flex items-center">
               <div>Link Bank Account</div>
               <div><span className="ml-3 material-icons">account_balance</span></div>
@@ -83,7 +68,7 @@ export function Dashboard({ }: any) {
             <h1>Dashboard</h1>
             <div>
               <Tooltip content={"Link/Update Account"} className="capitalize bg-gray-900 p-2">
-                <span onClick={() => setOpenProductSelection(!openProductSelection)} className="m-4 text-gray-600 hover:text-gray-200 cursor-pointer text-3xl material-icons">account_balance</span>
+                <span onClick={() => setOpenPlaidPrompt(!openPlaidPrompt)} className="m-4 text-gray-600 hover:text-gray-200 cursor-pointer text-3xl material-icons">account_balance</span>
               </Tooltip>
               <Tooltip content={"Unlink"} className="capitalize bg-gray-900 p-2">
                 <span onClick={() => unlinkBankAccount()} className="m-4 text-gray-600 hover:text-gray-200 cursor-pointer text-3xl material-icons">link_off</span>
@@ -104,86 +89,10 @@ export function Dashboard({ }: any) {
           </div>
         </>
       }
-
-      <Dialog size="lg" open={openProductSelection} handler={() => setOpenProductSelection(!openProductSelection)} className="bg-gray-800 max-h-screen overflow-auto">
-        <DialogHeader className="text-gray-100">Select Products</DialogHeader>
-        <DialogBody className="text-gray-200 flex flex-col">
-          <div className="grid grid-cols-3 gap-1">
-            {
-              [
-                "assets",
-                "auth",
-                "balance",
-                "identity",
-                "investments",
-                "liabilities",
-                "payment_initiation",
-                "identity_verification",
-                "transactions",
-                "credit_details",
-                "income",
-                "income_verification",
-                "deposit_switch",
-                "standing_orders",
-                "transfer",
-                "employment",
-                "recurring_transactions"
-              ].map((product) => {
-                return <div
-                  key={product}
-                  className={(selectedProducts.includes(product) ? "bg-sky-600" : "bg-zinc-800") + " hover:bg-zinc-500 hover:scale-105 shadow-xl cursor-pointer rounded-md p-2 text-sm font-bold"}
-                  onClick={() => {
-                    handleClickProduct(product);
-                  }}
-                >
-                  {product.replace(/_/g, " ").toUpperCase()}
-                </div>
-              })
-            }
-          </div>
-
-        </DialogBody>
-        <DialogFooter>
-          <Button
-            variant="text"
-            color="red"
-            onClick={() => {
-              setOpenProductSelection(!openProductSelection);
-              setSelectedProducts([]);
-              setLinkToken(null);
-            }}
-            className="mr-1"
-          >
-            <span>Cancel</span>
-          </Button>
-
-          {!linkToken ?
-            <Button variant="gradient" color="green" onClick={() => {
-              const formatProducts = selectedProducts.join(',');
-              // setOpenProductSelection(!openProductSelection);
-              getLinkToken({
-                variables: {
-                  products: formatProducts
-
-                },
-                onCompleted: (res) => {
-                  setLinkToken(res.createLinkToken['link_token']);
-                }
-              })
-            }}>
-              <span>Confirm</span>
-            </Button>
-            :
-            <PlaidLinkButton
-              linkToken={linkToken}
-              setLinkToken={setLinkToken}
-              selectedProducts={selectedProducts}
-              setSelectedProducts={setSelectedProducts}
-              setOpenProductSelection={setOpenProductSelection}
-            />
-          }
-        </DialogFooter>
-      </Dialog>
+      {/* Rerender everytime so we can access Plaid Link */}
+      {openPlaidPrompt &&
+        <PlaidLinkPrompt setOpenPlaidPrompt={setOpenPlaidPrompt} />
+      }
     </div>
   )
 }
